--- src/main/java/command/Command.java
@@ -1,7 +1,0 @@
-package command;
-
-import exception.DukeException;
-
-public interface Command {
-    public void execute(String input) throws DukeException;
-}
