--- src/main/java/data/task/Todo.java
@@ -1,8 +1,0 @@
-package data.task;
-
-public class Todo extends Task {
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
