--- src/main/java/data/task/builder/Builder.java
@@ -1,11 +1,0 @@
-package data.task.builder;
-
-import exception.DukeException;
-
-/**
- * Interface for all builders.
- * @param <T> The type of object to be built.
- */
-public interface Builder<T> {
-    public T buildFromString(String input) throws DukeException;
-}
