+++ src/main/java/duke/command/Command.java
@@ -1,0 +1,7 @@
+package duke.command;
+
+import duke.exception.DukeException;
+
+public interface Command {
+    public String execute(String input) throws DukeException;
+}
