+++ src/main/java/duke/data/task/Todo.java
@@ -1,0 +1,8 @@
+package duke.data.task;
+
+public class Todo extends Task {
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
