+++ src/main/java/duke/data/task/builder/Builder.java
@@ -1,0 +1,11 @@
+package duke.data.task.builder;
+
+import duke.exception.DukeException;
+
+/**
+ * Interface for all builders.
+ * @param <T> The type of object to be built.
+ */
+public interface Builder<T> {
+    public T buildFromString(String input) throws DukeException;
+}
