+++ src/main/java/duke/exception/InternalException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class InternalException extends DukeException {
+    public InternalException(String message) {
+        super("Internal exception: " + message);
+    }
+}
