+++ src/main/java/duke/exception/InvalidInputException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class InvalidInputException extends DukeException {
+    public InvalidInputException(String message) {
+        super("Invalid Input: " + message);
+    }
+}
