--- src/main/java/exception/DukeException.java
@@ -1,7 +1,0 @@
-package exception;
-
-public class DukeException extends Exception {
-    public DukeException(String message) {
-        super(message);
-    }
-}
