--- src/main/java/exception/InternalException.java
@@ -1,7 +1,0 @@
-package exception;
-
-public class InternalException extends DukeException {
-    public InternalException(String message) {
-        super("Internal exception: " + message);
-    }
-}
