--- src/main/java/exception/InvalidInputException.java
@@ -1,7 +1,0 @@
-package exception;
-
-public class InvalidInputException extends DukeException {
-    public InvalidInputException(String message) {
-        super("Invalid Input: " + message);
-    }
-}
